Changed various files to debug

+++ appinventor/appengine/src/com/google/appinventor/client/wizards/FileUploadWizard.java
@@ -83,49 +83,49 @@ public class FileUploadWizard extends Wizard {
     super(MESSAGES.fileUploadWizardCaption(), true, false);
 
     // Initialize UI
-    final FileUpload upload = new FileUpload();
-    upload.setName(ServerLayout.UPLOAD_FILE_FORM_ELEMENT);
-    setStylePrimaryName("ode-DialogBox");
-    VerticalPanel panel = new VerticalPanel();
-    panel.setVerticalAlignment(VerticalPanel.ALIGN_MIDDLE);
-    panel.add(upload);
-    addPage(panel);
+    final FileUpload upload = new FileUpload();                  //Makes an instance of FileUpload Widget called upload
+    upload.setName(ServerLayout.UPLOAD_FILE_FORM_ELEMENT);       //ServerLayout.UPLOAD_FILE_FORM_ELEMENT = "uploadFile", sets the uploadWidget name to be "uploadFile"
+    setStylePrimaryName("ode-DialogBox");                         
+    VerticalPanel panel = new VerticalPanel();                   //Makes an instace of the VerticalPanel called panel
+    panel.setVerticalAlignment(VerticalPanel.ALIGN_BOTTOM);      //Sets the panel to setVerticalAlignment to Align Middle
+    panel.add(upload);                                           //Adds the upload instance to the panel.
+    addPage(panel);                                              //Adds the panel to the page.
 
     // Create finish command (upload a file)
     initFinishCommand(new Command() {
       @Override
-      public void execute() {
-        String uploadFilename = upload.getFilename();
-        if (!uploadFilename.isEmpty()) {
-          final String filename = makeValidFilename(uploadFilename);
-          if(!TextValidators.isValidCharFilename(filename)){
-            createErrorDialog(MESSAGES.malformedFilenameTitle(), MESSAGES.malformedFilename(),
+      public void execute() {                                    //new command execute 
+        String uploadFilename = upload.getFilename();            //gets the Filename of the upload to the variable uploadFilename
+        if (!uploadFilename.isEmpty()) {                         //If this uploadFilename of the upload is not empty 
+          final String filename = makeValidFilename(uploadFilename);  //Calls the makeValidFilename with the uploadFilename returns a 
+          if(!TextValidators.isValidCharFilename(filename)){          //If TextValidators.isValidCharfilename returns False with filename passed, has single quotes or invalid characters,
+            createErrorDialog(MESSAGES.malformedFilenameTitle(), MESSAGES.malformedFilename(), //Error will come up with No Files Selected
               Error.NOFILESELECETED, folderNode, fileUploadedCallback);
-            return;
-          } else if (!TextValidators.isValidLengthFilename(filename)){
+            return;                                               //return nothing
+          } else if (!TextValidators.isValidLengthFilename(filename)){ //If TextValidators.isValidLengthFilename(filename) returns FALSE (has too long or too short of a file name)
             createErrorDialog(MESSAGES.filenameBadSizeTitle(), MESSAGES.filenameBadSize(),
-              Error.FILENAMEBADSIZE, folderNode, fileUploadedCallback);
-            return;
+              Error.FILENAMEBADSIZE, folderNode, fileUploadedCallback); //Error will come up with FileNameBadSize
+            return;                                               //returns nothing
           }
-          int nameLength = uploadFilename.length();
-          String fileEnd = uploadFilename.substring(nameLength-4, nameLength);
+          int nameLength = uploadFilename.length();               //takes the length of the uploadFilename and sets its to nameLength
+          String fileEnd = uploadFilename.substring(nameLength-4, nameLength); //Makes the String fileEnd equal to the file property suffix (.aia, .apk, etc)
 
-          if (".aia".equals(fileEnd.toLowerCase())) {
-            createErrorDialog(MESSAGES.aiaMediaAssetTitle(), MESSAGES.aiaMediaAsset(),
+          if (".aia".equals(fileEnd.toLowerCase())) {                                 //if the fileEnd equals to ".AIA", then another Error will come up with
+            createErrorDialog(MESSAGES.aiaMediaAssetTitle(), MESSAGES.aiaMediaAsset(), //AIAMEDIAASSET
               Error.AIAMEDIAASSET, folderNode, fileUploadedCallback);
-            return;
+            return; //returns nothing
           }
-          String fn = conflictingExistingFile(folderNode, filename);
-          if (fn != null && !confirmOverwrite(folderNode, fn, filename)) {
-            return;
+          String fn = conflictingExistingFile(folderNode, filename);  //Makes a string called fn and sets it equal to conflictExistingFile()
+          if (fn != null && !confirmOverwrite(folderNode, fn, filename)) { //if the fn is not null and user doesn't want to overwrite the file
+            return; //return nothing
           } else {
-            String fileId = folderNode.getFileId() + "/" + filename;
+            String fileId = folderNode.getFileId() + "/" + filename; //otherwise, set fileID to folderNODE.getFILEID()
             // We delete all the conflicting files.
-            for (ProjectNode child : folderNode.getChildren()) {
-              if (fileId.equalsIgnoreCase(child.getFileId()) && !fileId.equals(child.getFileId())) {
-                final ProjectNode node = child;
-                String filesToClose [] = { node.getFileId()};
-                Ode ode = Ode.getInstance();
+            for (ProjectNode child : folderNode.getChildren()) { //for each projectNode child in folerNode.getChildren()
+              if (fileId.equalsIgnoreCase(child.getFileId()) && !fileId.equals(child.getFileId())) { //if the fileID = child.GetFileID and the fileID doesn't equl child.getfileID
+                final ProjectNode node = child;                                                 //(if the fileID is equal to the child.getfileId without case and it doesnt equal child.getFileID considerinf case) (Different cases)
+                String filesToClose [] = { node.getFileId()};                                 //set the node equal to the child , creates an array filesToClose with that child node's file ID in it 
+                Ode ode = Ode.getInstance();                                            
                 ode.getEditorManager().closeFileEditors(node.getProjectId(), filesToClose);
                 ode.getProjectService().deleteFile(ode.getSessionId(),
                     node.getProjectId(), node.getFileId(),
@@ -142,37 +142,38 @@ public class FileUploadWizard extends Wizard {
               }
             }
           }
-          ErrorReporter.reportInfo(MESSAGES.fileUploadingMessage(filename));
+          ErrorReporter.reportInfo(MESSAGES.fileUploadingMessage(filename)); //Uploading the file to the server   @DefaultMessage("Uploading {0} to the App Inventor server")
 
           // Use the folderNode's project id and file id in the upload URL so that the file is
           // uploaded into that project and that folder in our back-end storage.
-          String uploadUrl = GWT.getModuleBaseURL() + ServerLayout.UPLOAD_SERVLET + "/" +
-              ServerLayout.UPLOAD_FILE + "/" + folderNode.getProjectId() + "/" +
+          String uploadUrl = GWT.getModuleBaseURL() + ServerLayout.UPLOAD_SERVLET + "/" + //uploadURL = urlprefix + upload + / + file
+              ServerLayout.UPLOAD_FILE + "/" + folderNode.getProjectId() + "/" + //+ / folderNode.getProjectId() + / + folderNode.getFileId() + "/" + filename
               folderNode.getFileId() + "/" + filename;
-          Uploader.getInstance().upload(upload, uploadUrl,
-              new OdeAsyncCallback<UploadResponse>(MESSAGES.fileUploadError()) {
+              ErrorReporter.reportInfo(uploadUrl);
+          Uploader.getInstance().upload(upload, uploadUrl,   //calls upload using upload widget, uploadURL created the line above
+              new OdeAsyncCallback<UploadResponse>(MESSAGES.fileUploadError()) { //returns with a error to wait later if it fails on callback
             @Override
-            public void onSuccess(UploadResponse uploadResponse) {
-              switch (uploadResponse.getStatus()) {
-              case SUCCESS:
-                ErrorReporter.hide();
+            public void onSuccess(UploadResponse uploadResponse) {  //on success for uploader
+              switch (uploadResponse.getStatus()) {    //gets the status of uploadResponse
+              case SUCCESS: 
+                ErrorReporter.hide(); //hides the erroReported
                 onUploadSuccess(folderNode, filename, uploadResponse.getModificationDate(),
-                    fileUploadedCallback);
+                    fileUploadedCallback); //updates modification date
                 break;
               case FILE_TOO_LARGE:
                 // The user can resolve the problem by
                 // uploading a smaller file.
-                ErrorReporter.reportInfo(MESSAGES.fileTooLargeError());
+                ErrorReporter.reportInfo(MESSAGES.fileTooLargeError()); //Error for file too Larger
                 break;
               default:
-                ErrorReporter.reportError(MESSAGES.fileUploadError());
+                ErrorReporter.reportError(MESSAGES.fileUploadError()); //Error for file didnt upload
                 break;
               }
             }
           });
         } else {
           createErrorDialog(MESSAGES.noFileSelectedTitle(), MESSAGES.noFileSelected(),
-              Error.NOFILESELECETED, folderNode, fileUploadedCallback);
+              Error.NOFILESELECETED, folderNode, fileUploadedCallback); //no files was selected
         }
       }
     });
@@ -199,12 +200,13 @@ public class FileUploadWizard extends Wizard {
     return filename;
   }
 
-  private String conflictingExistingFile(FolderNode folderNode, String filename) {
+  private String conflictingExistingFile(FolderNode folderNode, String filename) { 
     String fileId = folderNode.getFileId() + "/" + filename;
     for (ProjectNode child : folderNode.getChildren()) {
       if (fileId.equalsIgnoreCase(child.getFileId())) {
         // we want to return kitty.png rather than assets/kitty.png
-        return lastPathComponent(child.getFileId());
+        //return lastPathComponent(child.getFileId()); //see full path
+        return child.getFileId();
       }
     }
     return null;
@@ -221,7 +223,7 @@ public class FileUploadWizard extends Wizard {
 
   private void onUploadSuccess(final FolderNode folderNode, final String filename,
       long modificationDate, final FileUploadedCallback fileUploadedCallback) {
-    Ode.getInstance().updateModificationDate(folderNode.getProjectId(), modificationDate);
+    //Ode.getInstance().updateModificationDate(folderNode.getProjectId(), modificationDate);
     finishUpload(folderNode, filename, fileUploadedCallback);
   }
 
@@ -231,16 +233,28 @@ public class FileUploadWizard extends Wizard {
     FileNode uploadedFileNode;
     if (folderNode instanceof YoungAndroidAssetsFolder) {
       uploadedFileNode = new YoungAndroidAssetNode(filename, uploadedFileId);
+
     } else {
       uploadedFileNode = new FileNode(filename, uploadedFileId);
     }
+    Window.confirm("uploadedFileNode's getFileId(): " + uploadedFileNode.getFileId());
 
     Project project = Ode.getInstance().getProjectManager().getProject(folderNode);
-    uploadedFileNode = (FileNode) project.addNode(folderNode, uploadedFileNode);
 
-    if (fileUploadedCallback != null) {
-      fileUploadedCallback.onFileUploaded(folderNode, uploadedFileNode);
+
+    uploadedFileNode = (FileNode) project.addNode(folderNode, uploadedFileNode);
+    Window.confirm("uploadedFileNode's getFileId() after addNode " + uploadedFileNode.getFileId());
+
+    Iterable<ProjectNode> childrenNodes = project.getRootNode().getChildren();
+
+    for (ProjectNode node : childrenNodes) {
+        Window.confirm("getFullName: " + node.getFullName());
+        Window.confirm("getFileId: " + node.getFileId());
     }
+
+    /*if (fileUploadedCallback != null) {
+      fileUploadedCallback.onFileUploaded(folderNode, uploadedFileNode);
+    }*/
   }
 
   private void createErrorDialog(String title, String body, Error e,
